move SystemLoggingAO to separate class SystemManagementAO
